requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
idna==2.10
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
pytz==2021.1
requests==2.25.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
